requirements.txt
faiss-cpu
fastapi==0.109.2
langchain==0.3.16
langchain-community==0.3.16
langchain-core==0.3.33
langchain-openai==0.3.3
langchain-text-splitters==0.3.5
langgraph==0.2.73
pydantic==2.10.6
python-dotenv==1.0.1
uvicorn==0.34.0